Bug 716478 - update licence to MPL 2.
UltraBlame original commit: f4157e8c410708d76703f19e4dfb61859bfe32d8

+++ tools/l10n/l10n.py
@@ -4,39 +4,6 @@
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 import logging
 from optparse import OptionParser
 import os
@@ -54,7 +21,7 @@ def walk(base):
       pass
     yield (os.path.normpath(root[len(base)+1:]), files)
 
-def createLocalization(source, dest, apps = ['browser'], exceptions = {}):
+def createLocalization(source, dest, apps, exceptions = {}):
   '''
   Creates a new localization.
 
@@ -75,6 +42,9 @@ def createLocalization(source, dest, apps = ['browser'], exceptions = {}):
   assert os.path.isdir(source), "source directory is not a directory"
   clientmk = os.path.join(source,'client.mk')
   assert os.path.isfile(clientmk), "client.mk missing"
+
+  if not apps or not len(apps):
+    apps=['browser']
 
   if not os.path.isdir(dest):
     os.makedirs(dest)
@@ -124,7 +94,8 @@ def createLocalization(source, dest, apps = ['browser'], exceptions = {}):
         if ignore and ignore.search(f):
           
           continue
-        copy2(os.path.join(basepath, root, f), l10npath)
+        if not os.path.exists(os.path.join(l10npath,f)):
+          copy2(os.path.join(basepath, root, f), l10npath)
 
 if __name__ == '__main__':
   p = OptionParser()
@@ -132,6 +103,9 @@ if __name__ == '__main__':
                help='Mozilla sources')
   p.add_option('--dest', default = '../l10n',
                help='Localization target directory')
+  p.add_option('--app', action="append",
+               help='Create localization for this application ' + \
+               '(multiple applications allowed, default: browser)')
   p.add_option('-v', '--verbose', action="store_true", default=False,
                help='report debugging information')
   (opts, args) = p.parse_args()
@@ -144,4 +118,4 @@ if __name__ == '__main__':
                 {'searchplugins': '\\.xml$'},
                 'extensions/spellcheck': 'all'}
   createLocalization(opts.source, os.path.join(opts.dest, args[0]),
-                     exceptions=exceptions)
+                     opts.app, exceptions=exceptions)
